requirements.txt
# TTV Pipeline Dependencies
# Core dependencies
pillow>=10.0.0
pyttsx3>=2.90
numpy>=1.24.0

# Optional: gTTS fallback (requires internet)
# gtts>=2.3.0

# Testing (if needed)
# pytest>=7.0.0
